# Conflicts:
#	trunk/include/pdf/che_pdf_colorspace.h
#	trunk/include/pdf/che_pdf_contentbuilder.h
#	trunk/include/pdf/che_pdf_contentstring.h
#	trunk/include/pdf/che_pdf_gstate.h
#	trunk/include/pdf/che_pdf_objectstring.h
#	trunk/include/pdf/che_pdf_page_layout.h
#	trunk/projects/xcode/chelib.xcodeproj/project.pbxproj
#	trunk/src/pdf/che_pdf_contentbuilder.cpp
#	trunk/src/pdf/che_pdf_contentstring.cpp

+++ trunk/include/pdf/che_pdf_colorspace.h
@@ -27,7 +27,7 @@ public:
 	CHE_PDF_Color( CHE_Allocator * pAllocator = nullptr ) 
 		: CHE_Object( pAllocator ), mIndex(0)
 	{
-		for ( size_t i = 0; i < 4; ++i )
+		for ( uint32 i = 0; i < 4; ++i )
 		{
 			mComponent[i] = 0.0f;
 		}
@@ -67,7 +67,7 @@ public:
 		return mIndex;
 	}
 
-	FLOAT GetComponent( size_t index ) const
+	FLOAT GetComponent( uint32 index ) const
 	{
 		if ( index < mIndex )
 		{
